Added relation ship cascade

+++ bookapi.py
@@ -6,33 +6,43 @@ from app import db
 
 parser = reqparse.RequestParser()
 
+
 class BookListApi(Resource):
+
     def get(self):
         try:
             books = Book.query.all()
-            return jsonify([e.to_json() for e in books])
+            return jsonify([b.serialize() for b in books])
         except Exception as e:
             return str(e), 404
 
     def post(self):
-        parser.add_argument('name', type=str, help="Name of the book", required=True)
+        parser.add_argument('name', type=str,
+                            help="Name of the book",
+                            required=True)
+        parser.add_argument('editorial')
         parser.add_argument('published')
         args = parser.parse_args()
         name = args['name']
+        editorial = args['editorial']
         published = args['published']
         try:
-            book = Book(name=name, published=published, authors=None)
+            book = Book(name=name, editorial=editorial ,published=published)
             db.session.add(book)
             db.session.commit()
             return "Book added. book id={}".format(book.id), 201
         except Exception as e:
             return str(e), 400
 
+
 class BookApi(Resource):
+
     def get(self, id_book):
         try:
-           book = Book.query.filter_by(id=id_book).first()
-           return jsonify(book.serialize())
+            book = Book.query.filter_by(id=id_book).first()
+            if book is None:
+                return None, 404
+            return jsonify(book.serialize())
         except Exception as e:
             return str(e), 404
 
@@ -58,10 +68,11 @@ class BookApi(Resource):
 
     def delete(self, id_book):
         try:
-           book = Book.query.filter_by(id=id_book).first()
-           db.session.delete(book)
-           db.session.commit()
-           return jsonify(book.serialize())
+            book = Book.query.filter_by(id=id_book).first()
+            if book is None:
+                return None, 404
+            db.session.delete(book)
+            db.session.commit()
+            return jsonify(book.serialize())
         except Exception as e:
             return str(e), 404
-
